+++ etc/scripts/get-stack.sh
@@ -63,8 +63,10 @@ post_install_separator() {
 
 # determines the the CPU's instruction set
 get_isa() {
-  if arch | grep -q arm ; then
+  if arch | grep -q armv7 ; then
     echo arm
+  elif arch | grep -q aarch64 ; then
+    echo aarch64
   else
     echo x86
   fi
@@ -73,6 +75,11 @@ get_isa() {
 # exits with code 0 if arm ISA is detected as described above
 is_arm() {
   test "$(get_isa)" = arm
+}
+
+# exits with code 0 if aarch64 ISA is detected as described above
+is_aarch64() {
+  test "$(get_isa)" = aarch64
 }
 
 
@@ -145,10 +152,14 @@ do_ubuntu_install() {
     install_dependencies
     print_bindist_notice
     install_arm_binary
+  elif is_aarch64 ; then
+    install_dependencies
+    print_bindist_notice
+    install_aarch64_binary
   elif is_64_bit ; then
     install_dependencies
     print_bindist_notice
-    install_64bit_static_binary
+    install_64bit_standard_binary
   else
     install_dependencies
     print_bindist_notice
@@ -171,10 +182,14 @@ do_debian_install() {
     install_dependencies
     print_bindist_notice
     install_arm_binary
+  elif is_aarch64 ; then
+    install_dependencies
+    print_bindist_notice
+    install_aarch64_binary
   elif is_64_bit ; then
     install_dependencies
     print_bindist_notice
-    install_64bit_static_binary
+    install_64bit_standard_binary
   else
     install_dependencies
     print_bindist_notice
@@ -194,7 +209,7 @@ do_fedora_install() {
   if is_64_bit ; then
     install_dependencies "$1"
     print_bindist_notice
-    install_64bit_static_binary
+    install_64bit_standard_binary
   else
     install_dependencies "$1"
     print_bindist_notice
@@ -213,8 +228,16 @@ do_centos_install() {
 
   if is_64_bit ; then
     install_dependencies
-    print_bindist_notice
-    install_64bit_static_binary
+    case "$1" in
+      "6")
+        print_bindist_notice "libgmp4"
+        install_64bit_gmp4_linked_binary
+        ;;
+      *)
+        print_bindist_notice
+        install_64bit_standard_binary
+        ;;
+    esac
   else
     install_dependencies
     case "$1" in
@@ -263,7 +286,7 @@ do_alpine_install() {
   }
   install_dependencies
   if is_64_bit ; then
-    install_64bit_static_binary
+    die "Sorry, there is currently no 64-bit Alpine Linux binary available."
   else
     die "Sorry, there is currently no 32-bit Alpine Linux binary available."
   fi
@@ -277,11 +300,13 @@ do_sloppy_install() {
   info ""
 
   if is_arm ; then
-      install_arm_binary
+    install_arm_binary
+  elif is_aarch64 ; then
+    install_aarch64_binary
   elif is_64_bit ; then
-      install_64bit_static_binary
+    install_64bit_standard_binary
   else
-      install_32bit_standard_binary
+    install_32bit_standard_binary
   fi
   info "Since this installer doesn't support your Linux distribution,"
   info "there is no guarantee that 'stack' will work at all!  You may"
@@ -454,7 +479,7 @@ check_dl_tools() {
 
 # Download a Stack bindst and install it in /usr/local/bin/stack.
 install_from_bindist() {
-    IFB_URL="https://www.stackage.org/stack/$1"
+    IFB_URL="https://get.haskellstack.org/stable/$1.tar.gz"
     check_dl_tools
     make_temp_dir
 
@@ -498,12 +523,16 @@ install_32bit_standard_binary() {
   install_from_bindist "linux-i386"
 }
 
-install_64bit_static_binary() {
-  install_from_bindist "linux-x86_64-static"
+install_aarch64_binary() {
+  install_from_bindist "linux-aarch64"
 }
 
 install_32bit_gmp4_linked_binary() {
   install_from_bindist "linux-i386-gmp4"
+}
+
+install_64bit_gmp4_linked_binary() {
+  install_from_bindist "linux-x86_64-gmp4"
 }
 
 install_64bit_osx_binary() {